fix: cloudflare series pages
On cloudlflare pages, links to a series would 404. The links would take the plain text series name, but the series page router would convert to slug. This led to the page being lowercase but the links to it being capitalized which caused the 404. Netlify would route them through as case-insensitive i guess while cloudflare pages was case sensitive, leading to the issue.